pg_interface: move commit_reveal types to oko_types and ksn_interface
