+++ backend/oko_pg_interface/src/commit_reveal/index.ts
@@ -1,4 +1,3 @@
-export * from "./types";
 export { createSession } from "./insert_session";
 export { findByIdTokenHash, findBySessionId } from "./find_session";
 export { updateState } from "./update_session";
