+++ backend/oko_pg_interface/src/commit_reveal/insert_session.ts
@@ -1,6 +1,6 @@
 import type { Pool, PoolClient } from "pg";
 import type { Result } from "@oko-wallet/stdlib-js";
-import type { CommitRevealSession, CreateSessionParams } from "./types";
+import type { CommitRevealSession, CreateSessionParams } from "@oko-wallet/oko-types/commit_reveal";
 
 export async function createSession(
   db: Pool | PoolClient,
