+++ common/oko_types/package.json
@@ -28,7 +28,8 @@
     "./key_share_node_meta": "./dist/key_share_node_meta/index.js",
     "./tss_activate": "./dist/tss_activate/index.js",
     "./social_login": "./dist/social_login/index.js",
-    "./referral": "./dist/referral/index.js"
+    "./referral": "./dist/referral/index.js",
+    "./commit_reveal": "./dist/commit_reveal/index.js"
   },
   "dependencies": {
     "@oko-wallet/bytes": "^0.1.2-alpha.2",
