+++ key_share_node/ksn_interface/src/commit_reveal.ts
