+++ key_share_node/pg_interface/src/commit_reveal/find_session.ts
@@ -1,6 +1,6 @@
 import type { Pool, PoolClient } from "pg";
 import type { Result } from "@oko-wallet/stdlib-js";
-import type { CommitRevealSession } from "./types";
+import type { CommitRevealSession } from "@oko-wallet/ksn-interface/commit_reveal";
 
 export async function findByIdTokenHash(
   db: Pool | PoolClient,
