+++ key_share_node/pg_interface/src/commit_reveal/update_session.ts
@@ -1,6 +1,6 @@
 import type { Pool, PoolClient } from "pg";
 import type { Result } from "@oko-wallet/stdlib-js";
-import type { SessionState } from "./types";
+import type { SessionState } from "@oko-wallet/ksn-interface/commit_reveal";
 
 export async function updateState(
   db: Pool | PoolClient,
